\name{tgp-internal}
%% Part of the sensible export list but (currently?) documented as
%% internal (or waiting for documentation to be written):
%% <LIST>
\alias{tgp}
\alias{tree2c}
\alias{tgp.postprocess}
\alias{print.tgptraces}
\alias{tgp.read.XX.traces}
\alias{tgp.read.traces}
\alias{tgp.cleanup}
\alias{tgp.choose.as}
\alias{tgp.choose.center}
\alias{tgp.check.params}
\alias{tgp.partition}
\alias{tgp.get.trees}
\alias{tgp.plot.tree}
\alias{tgp.plot.parts.1d}
\alias{tgp.plot.parts.2d}
\alias{tgp.plot.slice}
\alias{tgp.plot.proj}
\alias{tgp.cands}
\alias{framify.X}
\alias{slice.interp}
\alias{slice.image}
\alias{slice.image.contour}
\alias{slice.contour}
\alias{slice.persp}
\alias{mean0.range1}
\alias{undo.mean0.range1}
\alias{check.matrix}
\alias{mr.checkrez}
\alias{check.itemps}
\alias{check.slice}
\alias{check.proj}
\alias{check.sens}
\alias{getlocs}
\alias{print.tgp}
\alias{sens.plot}
\alias{mr.plot}
\alias{mr.checkres}
%% </LIST>
%% Currently (?) exported as used in the grid.layout.Rd \example ...
%% <LIST>
\alias{layout.torture}
%% </LIST>

\title{Internal Treed Gaussian Process Model Functions}
\description{
  Internal Treed Gaussian Process Model functions
}
\details{
  These are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written :)).
}
\references{
\url{http://www.ams.ucsc.edu/~rbgramacy/tgp.html}
}
\keyword{ internal }
